requirements-dev.txt
# Install base requirements
-r requirements.txt

# Testing
pytest-mock==3.11.1
pytest-sugar==0.9.7  # For better test output
pytest-randomly==3.13.0  # For random test ordering
pytest-timeout==2.1.0  # For test timeouts
coverage==7.3.0

# Debugging
ipython==8.14.0
ipdb==0.13.13
line-profiler==4.1.1

# Code quality
pre-commit==3.3.3
pylint==2.17.5
bandit==1.7.5  # For security checks
# safety==2.3.5  # For dependency security checks
vulture==2.9.1  # For dead code detection

# Type checking
mypy-extensions==1.0.0
# types-all==1.0.0

# Documentation
mkdocs==1.5.2
mkdocs-material==9.2.3
mkdocstrings==0.22.0
mkdocstrings-python==1.5.2
sphinx<7.0.0  # Pinned to be compatible with sphinx-rtd-theme
sphinx-rtd-theme==1.2.2

# Development tools
bump2version==1.0.1  # For version management
twine==4.0.2  # For PyPI uploads 